Rename files for object format parsing.
