+++ framework/src/bundle/BundleObjFile.cpp
@@ -19,7 +19,7 @@
 
 =============================================================================*/
 
-#include "ModuleObjFile.h"
+#include "BundleObjFile.h"
 
 #include <cstring>
