+++ framework/src/bundle/BundleObjFile.h
