+++ framework/src/bundle/BundlePEFile.cpp
@@ -19,7 +19,7 @@
 
 =============================================================================*/
 
-#include "ModuleObjFile.h"
+#include "BundleObjFile.h"
 
 #include "cppmicroservices_pe.h"
